refactor: :truck: rinomina cartella
rinomino cartella da screen a pages
